Dotnet21

+++ DemoTests/AdditionTests.cs
@@ -27,7 +27,7 @@ namespace Demo.Tests
         public void EmptyListTest()
         {
             Addition a = new Addition();
-            List<int> values = null;
+            List<float> values = null;
             var result = a.Add(values);
             Assert.AreEqual(0, result);
         }
@@ -37,34 +37,35 @@ namespace Demo.Tests
         public void EmptyListTest()
         {
             Addition a = new Addition();
-            List<int> values = null;
+            List<float> values = null;
             _ = a.Add(values);
         }
 
+        /*
         [TestMethod]
         public void OutofBoundsTest()
         {
             // Arrange
-            int lessThanMax = Int32.MaxValue - 5;
-            List<int> values = new List<int>() { lessThanMax, 3, 3 };
+            float lessThanMax = float32.MaxValue - 5;
+            List<float> values = new List<float>() { lessThanMax, 3, 3 };
 
             // Act
             Addition a = new Addition();
-            int sum = a.Add(values);
+            float sum = a.Add(values);
 
             // Assert
-            Assert.IsTrue(sum > Int32.MaxValue);
+            Assert.IsTrue(sum > float32.MaxValue);
         }
-
+        */
         [TestMethod]
         public void IsResultCorrect()
         {
             // Arrange
-            List<int> values = new List<int>() { 2, 3 };
+            List<float> values = new List<float>() { 2, 3 };
 
             // Act
             Addition a = new Addition();
-            int sum = a.Add(values);
+            float sum = a.Add(values);
 
             // Assert
             Assert.IsTrue(sum == 5);
